requirements.txt
playwright>=1.47.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
ollama>=0.1.7
pydantic>=2.7.0
python-dotenv>=1.0.1
tenacity>=9.0.0
tqdm>=4.66.0
groq
python-dotenv